backend/requirements.txt
Django==4.2.1
djangorestframework==3.14.0
django-cors-headers==4.1.0
Pillow==9.5.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
django-filter==23.2
djangorestframework-simplejwt==5.2.2
stripe==5.4.0
gunicorn==20.1.0
whitenoise==6.4.0
django-storages==1.13.2
boto3==1.26.137
redis==5.0.1
elasticsearch==8.11.0
django-elasticsearch-dsl==7.2.4
pytest==7.4.3
pytest-django==4.7.0 